+++ packages/libs/client-examples/src/example-contract/util/client.ts
@@ -1,4 +1,5 @@
-import { ITransactionDescriptor, createClient } from '@kadena/client';
+import type { ITransactionDescriptor } from '@kadena/client';
+import { createClient } from '@kadena/client';
 import type { ChainId, ICommand, IUnsignedCommand } from '@kadena/types';
 
 // you can edit this function if you want to use different network like dev-net or a private net
